ENG-1814 Use 4 bytes for int32, int16 and int8 instead of 8 bytes.
Summary:
We currently use fixed 8 bytes for all types of ints (int8, int16, int32 and int64). As a
result, for smaller int types we are wasting a lot of space in memory and on disk. With this diff,
we ensure int8, int16 and int32 use 4 bytes. I'll follow up this diff with another diff to change
float to 4 bytes.

Note that this is a format change.

Test Plan: Enhanced unit tests for int32, jenkins

Reviewers: mikhail, akashnil, hector

Reviewed By: hector

Subscribers: hector, ybase, kannan, karthik

Differential Revision: https://phabricator.dev.yugabyte.com/D2503