ENG-1690 Increase rocksdb initial sequence number to 2^50
Summary:
The initial sequence number of 2^40 was probably cutting it too close for some potentially
large usecases. If we have 2^40 sequence numbers for bulk loads, we have approximiately 10^12
rocksdb keys. Since for CQL each rocksdb key is roughly a column, if we have 100 columns we would be
able to support 10 billion rows per tablet. Although, that is a large number its not impossible for
us to have such a usecase. Using 2^50 gives us enough room for bulk loads.

Test Plan: jenkins

Reviewers: sergei, mikhail, karthik, kannan

Reviewed By: kannan

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2500