[ENG-1824] Comparable Encoding for Floats and Doubles
Summary:
We want to encode Floats as 4 byte strings and doubles as 8 byte strings so that x < y if and only if Enc(x) < Enc(y).

Notice that  a float is represented in memory as - [sign bit S] [8-bit unsigned int E] [23 bit unsigned int M]
The value of the float is (-1)^S * 2^(E - 127) * (1 + M * 2^(-23))
https://en.wikipedia.org/wiki/Single-precision_floating-point_format

If S is 0, the number is positive. In that case, the representation is already comparable, because f(E, M) = 2^(E - 127) * (1 + M * 2^(-23)) is an increasing function of the pair (E, M)

If E1 > E2 then f(E1, M1) > f(E2, M2) and if M1 > M2, then f(E, M1) > f(E, M2).

To support negative numbers, the first thing we need is to flip the sign bit, so that negative numbers have first bit 0, and positive numbers have first bit 1.

This is almost done, but there's one issue that two negative numbers don't compare correctly, because [E][M] part of the number compares in the order of their absolute values, which is the opposite of the real values of the float. So we need to invert (bitwise negation) everything, if the number is negative.

Ultimately the logic is very simple - consider the float v as a 32-bit unsigned int (by using reinterpret cast for example):

  - For positive numbers Enc(v) = v ^ 1<<31. (Flip the sign bit from 0 to 1)
  - For negative numbers, Enc(v) = ~v (bitwise negation of all bits)

Note that 0 and -0 are different for floating point numbers (0x00000000 and 0x80000000 respectively), their respective encodings are (0x80000000 and 0x7FFFFFFF respectively) which compare in the same way Cassandra does.

Test Plan: Added unit tests

Reviewers: robert, rahuldesirazu, pritam.damania, mikhail, yql, hector

Reviewed By: hector

Differential Revision: https://phabricator.dev.yugabyte.com/D2510