ENG-1833 On Prem Edit Universe, Not Enough Nodes configured status when there are enough nodes
Summary:
During Edit Universe OnPrem

During OnPrem Edit,  [As in JIRA]
We show the error message when number of nodes > number of nodes configured.
To calculate this in Edit, we also need to include the number of nodes in the existing OnPrem universe.

https://drive.google.com/a/yugabyte.com/file/d/0B2MFlVNRpzl3SnRxTi1SajgzV0U/view?usp=sharing

Test Plan:
Go to Edit Universe for OnPrem.
The error message should be shown only when numNodes > numNodesConfigured

Reviewers: jeff, ram

Reviewed By: ram

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D2525