ENG-1863: Status endpoint for customer tasks should show status of subtaskgroups
Summary: As in the title. Commissioner now goes straight to the task_info corresponding to a CustomerTask and grabs the information it needs from that task_info. This deduplicates some code, reduces the number of layers of code we have to go through to get the information, and removes dependencies on in-memory objects that are lost once a CustomerTask has finished (whether in success or failure).

Test Plan: sbt test; hit endpoint through REST api

Reviewers: pritam, ram, bharat

Reviewed By: ram, bharat

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D2568