Allow concurrent compactions in Level 0 of rocksdb.
Summary:
This commit is basically a revert of
https://github.com/facebook/rocksdb/commit/2ae4d7d70864ceadc9668f1b7e47d26d839a8f1a. In YB we only
use 1 level and have universal compaction and as a result we should support concurrent compactions
in level 0. It seems like this is an issue for rocksdb given that subsequent levels can't have
overlapping ranges and hence concurrent compactions in level 0 might cause issues as the files are
pushed down to subsequent levels.

Test Plan: jenkins

Reviewers: mikhail, timur, sergei

Reviewed By: sergei

Subscribers: kannan, bogdan, amitanand, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2571