ENG-1874: Fixing data race issues with accessing handles returned from the block cache.
Summary: Fixing a data race bug found from the TSAN tests that was reported in ENG-1874. The problem is that LRULookup has a mutex that changes the query ID of the LRUHandle, but there are external functions also accessing the query id of the same handle that was returned previously. The fix is to move the statistics reporting into the cache itself to avoid the issue of having to call external functions to access the query id.

Test Plan: TSAN tests

Reviewers: karthik, bogdan, kannan, sergei, mikhail, sean

Reviewed By: mikhail, sean

Subscribers: timur, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2576