ENG-1889: Fix TestPrepareExecute ASAN issue
Summary:
The heap-use-after-free happens because when a prepared statement fails to execute and is considered stale, it is deleted from `prepared_stmts_map_` in `CQLService` and `stmt_` in `CQLProcessor` in the StatementExecutedCallback. There are still references to the statement in the call stack before the call finishes and returns from `CQLProcessor::ProcessExecute()` and the heap-use-after-free results because the statement is purged. We should keep another `shared_ptr` reference to the statement in `CQLProcessor::ProcessExecute()` to keep it in memory until the method returns. Same case for `CQLProcessor::ProcessBatch()`

This diff also fixes ENG-1893 to return keyspace-not-found error properly as `ErrorCode::KEYSPACE_NOT_FOUND` so that TestPrepareExecute.testDDL() will receive the proper error status.

Test Plan: Jenkins tests

Reviewers: bogdan, oleg

Reviewed By: oleg

Subscribers: mihnea, yql

Differential Revision: https://phabricator.dev.yugabyte.com/D2602