Handle Server Side Error object from Prometheus API response
Summary:
Currently, if the metrics response returns an error object, the UI will crash, because we do not handle metrics.error objects, when metrics.data is undefined. For eg. when Prometheus server crashes.

{F4048}

An error object is different from an empty metrics.data object.

We do not have a top level error object for metrics , to show in the UI what the error is , (eg. Prometheus server may have crashed, leading to java.new.connectionRefusedException)

But for now, if metrics.data is undefined and metrics.error , we prevent the graph UI from crashing, by not showing any graph data.

{F4047}

The other option may be to show individual error objects for each graph. (Since we do not have a top level error object)
{F4049}

Test Plan:
If prometheus sever crashed, do not show any metrics or graph placeholders.

Reviewers: bogdan, jeff, ram

Reviewed By: ram

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D2467