ENG-1938 PlacementInfo contains stale data if only userIntent has changed
Summary:
Update PlacementInfo from the nodeDetailsSet when the payload hits the Create/Update endpoint.
The nodeDetailsSet is the source of truth and placementInfo should be updated with that value.

Test Plan:
Create a 1 node RF 1 cluster.
The Universe Detail map should render the correct AZ.
The AZ list should render the correct AZ.

Reviewers: ram, jeff, bharat

Reviewed By: bharat

Subscribers: jenkins-bot, bogdan, sid, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D2624