ENG-1584 Ensure YBClient uses placement information to determine closest replicas.
Summary:
In D2249, we ensured that RemoteTabletServer includes the placement information for a
tserver. In this diff, we ensure that when we use SelectTserver() with CLOSEST_REPLICA, we look at
the placement_zone and placement_region of the tserver and compare that to the client to pick the
closest replica.

Test Plan:
1) jenkins
2) Unit tests that simulate different regions and verify the appropriate behavior.

Reviewers: hector, robert, bharat

Reviewed By: bharat

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2649