ENG-1964: Ensure RF changes during create universe works end to end.
Summary:
Remove the requirement to have masters pre-selected during configure steps for create. The set of masters will be picked by create task after submit.

This also remove the need remove or transfer masters as nodes are changed on the placement AZ's (right side of the create page).

Also there was a bug that we were re-picking a node that was already a master (after RF change). Fixed that too by removing a node marked with master from potential victim set.

Note that reducing the value of RF to 1 and then to say 5, is not picking more AZ's. This will be handled in followups. Might need UI changes too - FYI @pritam.

Test Plan:
From UI, in create page, changed RF from 3 to 5 and hit submit - ensured there are 5 masters during task run. Earlier it used to create 3 masters when RF=5 was chosen.

```
2017-08-10 12:20:51,449 [INFO] from com.yugabyte.yw.commissioner.tasks.CreateUniverse in TaskPool-0 - Started CreateUniverse(4956240c-68ce-4b27-9c1f-2faf0b38a251) task.
2017-08-10 12:20:51,454 [DEBUG] from com.yugabyte.yw.models.Universe in TaskPool-0 - Swapped universe 4956240c-68ce-4b27-9c1f-2faf0b38a251:bsdfs details to [{"deviceInfo":null,"cloud":null,"universeUUID":null,"expectedUniverseVersion":0,"userIntent":null,"nodePrefix":null,"nodeDetailsSet":[],"placementInfo":null,"updateInProgress":true,"updateSucceeded":false}] with new version = 2.
2017-08-10 12:20:51,457 [DEBUG] from com.yugabyte.yw.commissioner.tasks.UniverseTaskBase in TaskPool-0 - Locked universe 4956240c-68ce-4b27-9c1f-2faf0b38a251 at version -1.
2017-08-10 12:20:51,457 [INFO] from com.yugabyte.yw.common.PlacementInfoUtil in TaskPool-0 - Subnet map has 3, nodesMap has 5, need 5 masters.
2017-08-10 12:20:51,457 [INFO] from com.yugabyte.yw.common.PlacementInfoUtil in TaskPool-0 - Chose node 'yb-1-bsdfs-fake-n2' as a master from subnet subnet-4e46ca07.
2017-08-10 12:20:51,457 [INFO] from com.yugabyte.yw.common.PlacementInfoUtil in TaskPool-0 - Chose node 'yb-1-bsdfs-fake-n1' as a master from subnet subnet-559ff532.
2017-08-10 12:20:51,457 [INFO] from com.yugabyte.yw.common.PlacementInfoUtil in TaskPool-0 - Chose node 'yb-1-bsdfs-fake-n3' as a master from subnet subnet-cc32d797.
2017-08-10 12:20:51,457 [INFO] from com.yugabyte.yw.common.PlacementInfoUtil in TaskPool-0 - Chose node 'yb-1-bsdfs-fake-n4' as a master from subnet subnet-4e46ca07.
2017-08-10 12:20:51,457 [INFO] from com.yugabyte.yw.common.PlacementInfoUtil in TaskPool-0 - Chose node 'yb-1-bsdfs-fake-n5' as a master from subnet subnet-cc32d797.
```

Reviewers: ram, pritam, jeff

Reviewed By: jeff

Subscribers: jenkins-bot, pritam, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D2656