ENG-1727 : Fix log cache EvictSomeUnlocked issue
Summary: Consensus peer corresponding to dead tablet peer was keeping a shared_ptr to replicate msgs even after it knew the tablet was not reachable. This meant log cache couldn't evict these messages, and so it filled up quickly (and CPU time would be wasted on EvictSomeUnlocked which linearly scans all entries up till a certain OpID to see which ones can be evicted).

Test Plan: Working on adding a test.

Reviewers: amitanand, bogdan

Reviewed By: bogdan

Subscribers: kannan, karthik, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2578