Supporing Intents and their encoding/decoding
Summary:
This is a remade version of D2242

Intent is a tuple (SubDocKey, IntentType, TransactionId, Value) [Value may be optional, IntentType incorporates the isolation level]

When writing an intent, we put the (SubDocKey, IntentType, TransactionId) into the rocksdb key, and (Value) into the rocksdb value. This may change in future if we need to add more fields to intents.

Intent keys also start with the GroupEnd Character, to ensure that they go to a separate part of the keyspace in the same rocksdb instance.

Test Plan: Added a unit test for encoding decoding

Reviewers: timur, sergei, mikhail

Reviewed By: mikhail

Subscribers: timur, eng

Differential Revision: https://phabricator.dev.yugabyte.com/D2554