ENG-1828 Show Number of Nodes in Universe Detail page
Summary:
We need to show the number of nodes in the Universe Detail page.

Modified UniverseResources component to take in numNodes object when renderMode is Display.

Render Mode can either be Configure or Display.

{F4433}

Test Plan:
Go to Universe Detail page of a newly created universe.

You should be able to see the numNodes for that universe.

Reviewers: jeff, ram

Reviewed By: ram

Subscribers: jenkins-bot, sid, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D2517