Fix inet test failures on GCP due to lack of ipv6 support.
Summary:
By default building an InetAddress object from a string would try to resolve the provided
string as a hostname. This used getaddrinfo() underneath, which would skip the DNS lookup if it
found a valid IP address notation. Although for IPv6, this wouldn't work if the machine wasn't
configured to use IPv6 since getaddrinfo() couldn't reach the provided IP and didn't find a route
for it either.

The solution was to check if we have a valid IP addr string and resolve the host only if it wasn't a valid IP addr string.

Test Plan: jenkins

Reviewers: mikhail

Reviewed By: mikhail

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2698