Change Provider map icon to custom icon type
Summary:
Earlier we were using the RootAZ data marker for Region Map as well.
Changed that to custom type.
Show number of AZs in the region inside the circle.

{F4435}

{F4436}

Test Plan:
Go to Provider config.
You should see the provider region markers.
The html on the marker should be the number of AZs in the marker

Reviewers: jeff, ram

Reviewed By: ram

Subscribers: sid, jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D2594