ENG-1957 Fix tsan issues in FaultInjectionTest.
Summary:
The `sleeping_task_low` variable was being passed to a background thread, but we were not
waiting for the thread to finish before returning from the method. This would cause a background
thread to hold onto pointer to an object which when destroyed caused tsan issues.

Test Plan: run the tests 100 times.

Reviewers: mikhail

Reviewed By: mikhail

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2711