Fix UI warnings from BulkLoad.js
Summary: I was using `!= undefined` to catch when something was null and undefined. Now using lodash instead to check that the instanceCount, if specified, is a number and is a number > 0. Also added in a radix for the `parseInt` conversion since that can apparently cause occasional issues.

Test Plan: npm start, verify no warnings

Reviewers: pritam, ram, shawn

Reviewed By: shawn

Subscribers: jenkins-bot, ui

Differential Revision: https://phabricator.dev.yugabyte.com/D2724