ENG-1651 Fix ASAN issues due to leak in peer manager.
Summary:
We were storing naked pointers in the peers_ map despite the fact that it was supposed to
own the pointers. As a result when we erased items in the map in a few places, we forgot to delete
the appropriate pointers. I've changed this map to store unique_ptrs so that memory management is
much simpler.

Test Plan:
1) run ClientTest.TestReadFromFollower 100 times (which was failing due to this)
2) jenkins

Reviewers: mikhail, hector

Reviewed By: hector

Subscribers: bogdan, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2733