ENG-1990 Have ability to delete nodes that couldn't be provisioned (UI Part)
Summary:
Create UI action to delete node which couldn't be provisioned and remained in the "ToBeAdded" state.

Since Delete Node from DB is now an async task, add timeOut to response handler to fetch response.

Test Plan:
Try to delete a node that couldn't be provisioned.
The delete must be successful and the the userIntent, nodeDetailsSet and placementInfo should be updated accordingly.

Reviewers: ram, jeff

Reviewed By: jeff

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D2696