ENG-2012 Add a flag to disable ntp clock synchronization errors.
Summary:
We shouldn't fatal if there are large clock synchronization errors in YB. This is only an
issue for distributed transactions and we can address this in the long term for distributed
transactions. Currently, there is no reason for us to fail if the ntp clocks are not synchronized.

Test Plan: jenkins

Reviewers: akashnil, mikhail, bogdan

Reviewed By: mikhail

Subscribers: bogdan, ybase, kannan, karthik

Differential Revision: https://phabricator.dev.yugabyte.com/D2773