ENG-2023 Error icon being showed everytime a task is being performed
Summary:
In current code, the else block sets the status to "error" condition.
Therefore while task is in flight, it comes to the else block and intermittently comes to the error condition.

Fixed this by adding an unknown "loading" condition, which is the status state while task is in progress.
https://drive.google.com/a/yugabyte.com/file/d/0B2MFlVNRpzl3N1hrdmh5UndSWVk/view?usp=sharing

Test Plan:
Create a new task.
Refresh the page, the status should so loading instead of error while task fetch is in progress.

Reviewers: ram, jeff

Reviewed By: jeff

Subscribers: hector, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D2766