ENG-1652: Adding an endpoint that can write out Prometheus style metrics
Summary:
Exposing an endpoint for Prometheus to pull data. Formatting the data as per Prometheus
documentation. There's plenty of follow-up to do, but I want to get something in and start testing
on portal and slowly roll metrics over until we can get rid of collectd all together!

Test Plan:
ran prometheus on my dev-server, setup a scrape config to the tserver/master ports and
setup the path to /prometheus-metrics and then confirmed the graphs show up!

https://phabricator.dev.yugabyte.com/F5519

Reviewers: mikhail, ram, karthik, jeff, bharat

Reviewed By: bharat

Subscribers: devops, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2797