Fix issue with User-Defined Types being allowed in other keyspaces
Summary:
In CQL UDTs can only be used in the same keyspace they are defined in.
Therefore, the UDT names don't include the keyspace because that is implied from the context.
Before, we included the keyspace name in the system_schema tables (columns and types) to disambiguate.
This confused the Cassandra connector in some cases.

Test Plan:
updated java TestUserDefinedTypes
for name:
```
 create keyspace test;
 create type test.udt (a int, b int);
 create table test.tb (a int primary key, b test.udt, c frozen<test.udt>);
 describe table test.tb;
```
Should return udt type name without the keyspace.

Reviewers: robert, neil, pritam.damania

Reviewed By: pritam.damania

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D2804