Insert default extra flags in the beginning of master/tserver command line so that they can be overridden
Summary:
Insert default extra flags in the beginning of master/tserver command line so that they can be
overridden This is to fix RaftConsensusITest.TestEarlyCommitDespiteMemoryPressure. Without this fix,
we're adding the 1G default memory limit after the custom memory limit specified by the test.
It looks like I broke this test in D2778.

Test Plan: Jenkins

Reviewers: amitanand, bogdan, hector

Reviewed By: hector

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2847