Rename TaskList to SubTaskGroup
Summary:
Have gotten feedback from a few people that "TaskList" is a confusing name that implies a certain ordering and sequentiality to the tasks it contains. Also, it seems like it should contain "Tasks" which are a generic thing.

In reality, it only contains SubTasks (as in, tasks which must have a parent). Also, the purpose of this object it to group together subtasks that can execute concurrently. Thus "SubTaskGroup" seems to be a much more appropriate name.

Test Plan: sbt test

Reviewers: karthik, bharat

Reviewed By: bharat

Subscribers: jenkins-bot, mihnea, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D2860