Fix TSAN Failure for ybsql static column test
Summary:
Tests assumed last paging result is always empty, but if this is last tablet it could
return empty page already on the last real result.
This is because of SetPagingStateIfNecessary in DocRowwiseIterator checks if there is a
next key to read.
TSAN tests use 2 tablets instead of 8 increasing the chance of this issue being hit.

Test Plan: jenkins

Reviewers: robert, mikhail

Reviewed By: mikhail

Subscribers: mikhail, yql

Differential Revision: https://phabricator.dev.yugabyte.com/D2867