YBClientService now allows multiple clients to exist
Summary: We were running into a concurrency bug where only one client could exist at a time and, occasionally, that one client would get overwritten while still in use. Fix is to simply allow the creation of multiple clients and provide a simple utility function to close the client when finished with it. Whoever requests a client must be sure to close the client when finished with it.

Test Plan:
- sbt test
- Create universe 1. Once finished, create universe 2. While universe 2 is being created, continuously query the listTables endpoint on universe 1 (this will request a new client while another one is being used). Verify universe 2 is still successfully created.

Reviewers: bogdan, bharat

Reviewed By: bharat

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D2868