ENG-1652: Migrating dashboards from collectd to Prometheus
Summary:
Converting system specific metrics from psutils style collectd output to
using the node_exporter metrics. Had to change quite a bit of the mechanics to allow for this
change, as apparently Prometheus does not keep the __name__ label when doing rates, as that's
apparently an anti-pattern, but we were previously using the type label for keeping the name
essentially. Migrated this to be essentially multiple queries and merged on the client side.

Test Plan: Testing with a local YW pointing to my prometheus and a custom YB universe.

Reviewers: karthik, jeff, bharat, pritam, ram

Reviewed By: ram

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D2887