Use a higher timeout for tserver_yb_client_default_timeout_ms for local_cluster_ctl.sh
Summary:
Sometimes the masters take a while to load the syscatalog data from disk and the tservers
timeout based on the 5 second timeout before a master leader is elected and syscatalog is loaded
successfully. Increasing this timeout for local_cluster_ctl.sh so that we don't have this
non-deterministic behavior.

Test Plan: Run the script.

Reviewers: mikhail, bharat, bogdan, venkatesh

Reviewed By: venkatesh

Subscribers: venkatesh, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2901