Silence debug message for system tables in PartitionMetadata.
Summary: Silence noisy debug messages in PartitionMetadata when it cannot find host for a system table. This is expected when the masters run on separate IP addresses than tservers which is the case for our Jenkins tests.

Test Plan: Run a long-running CQL Jenkins tests and verify the debug messages for system tables do not appear anymore.

Reviewers: pritam.damania

Reviewed By: pritam.damania

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D2900