ENG-2105: Make CloudBootstrap work for GCP
Summary: CloudBootstrap, which is a way of putting the cloud configuration in the task framework as a single task, only supported AWS. This diff adds support for GCP to CloudBootstrap. Also created a cloud-agnostic AbstractInitializer as part of this in order to move cloud configuration logic closer to a cloud-agnostic approach.

Test Plan:
- sbt test
- delete GCP from database; hit `http://localhost:9000/api/customers/<cUUID>/providers/<pUUID>/bootstrap` endpoint with payload `{"regionList": ["us-west1"]}`; verify GCP is properly bootstrapped

Reviewers: bogdan, bharat

Reviewed By: bharat

Subscribers: jenkins-bot, ram, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D2902