ENG-2097: Leader stepdown infinite loop prevention
Summary:
When the load balancer sends a leader stepdown request to a tablet server, and a similar request has
recently failed (as identified by the tablet id and intended leader), we reject the stepdown request
for a configured time period (e.g. 20 seconds or so). Otherwise it is possible that the load
balancer will continue retrying the same stepdown too soon, and by doing so prevent the intended
leader from ever catching and cause extended unavailability of the tablet.

Test Plan: Jenkins

Reviewers: hector, robert, bogdan, sergei

Reviewed By: sergei

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2895