ENG-2104: Cleaning up flags and changing data/wal dir structure
Summary:
As per discussions, we want to move towards the server actually creating the directories
necessary including a toplevel `yb-data` in which we would drop a `master` and a `tserver` dir.
This would still require the user to only provide the root in which to create this structure.

All the dirs in `fs_data_dirs` are treated as roots, under which we would put
`/yb-data/{master,tserver}/data`
If we do not specify `fs_wal_dirs`, we take `fs_data_dirs` as the set of roots under which to create
our directory structures and add a `wals` as a peer to the `data` dir.
If we do not specify a `log_dir`, we would create a subsequent `logs` dir as a peer to `data` and `wals`,
down the path from the first dir `fs_data_dirs`.

Additionally this script does the following:
- rename flag `default_num_replicas` to `replication_factor`
- rename flag `raft_get_node_instance_timeout_ms` to `master_discovery_timeout_ms`
- update flag `master_discovery_timeout_ms` to allow for 1h for masters to sync up
- rename binary `bin/yugabyte-cli` to `bin/yb-ctl` and change the relevant manifest

Test Plan: tested with `bin/yb-ctl create/add_node/destroy`

Reviewers: karthik, sid, mikhail, kannan

Reviewed By: kannan

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2916