ENG-2104 subtask: default webserver_doc_root automatically
Summary:
If webserver_doc_root was not specified, it was being determined as $YB_HOME/www.

With this change, if YB_HOME is also not specified, we  automatically determine "www" location based on how release packaging is done. We assume the "bin" and "www" directories are at the same level. So, if binary path is /opt/yugabyte/tserver/bin/yb-tserver, the gflag for webserver_doc_root will be automatically set to /opt/yugabyte/tserver/www.

Test Plan: unit tests, and verify in local cluster

Reviewers: bogdan, kannan, mikhail

Reviewed By: mikhail

Subscribers: ybase, sid

Differential Revision: https://phabricator.dev.yugabyte.com/D2931