Reset Node Name on Universe Delete
Summary:
When deleting an on-prem universe.
We should reset the nodename to "" (empty string)
So that the node can be reused to re-create a universe with same name.

Test Plan:
Create onprem universe with name "onprem"
Delete universe.
Try to create universe with same name "onprem"
Universe create should succeed.

Reviewers: jeff, ram

Reviewed By: ram

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D2973