Improve stability of TestConsistencyLevels.
Summary:
This test used to wait for all followers to catch up by performing a bunch of reads to
ensure it hit all followers. Although TestUtils.waitFor(), would check if the given condition has
evaluated to true and exit its loop and then check the condition again. In our case this second
check to hit a follower which hasn't caught up yet. As a result, to fix this we ensure that the
condition only evaluates to true after we have tested it 10 times.

Test Plan: run the unit test multiple times.

Reviewers: hector, robert

Reviewed By: robert

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D2986