[ENG-2157] Optimize selects with IN condition on partition columns
Summary:
Previously such queries lead to a full table scan and in-memory filtering. Now handling is
similar to Cassandra. Selects with IN on the partition columns translate to several read requests
internally, one for each possible partition key. Limit and paging state work as expected.

The lists of values from the IN conditions are converted to ordered sets (remove duplicates,
order in ascending order) like in Cassandra. Therefore the results are ordered in ascending
order of hash columns.

Test Plan: jenkins, ql-query-test in C++ and TestSelect/TestBindVariable in java.

Reviewers: neil, robert

Reviewed By: robert

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D3109