ENG-2136: Increase timeout for Cassandra drop table.
Summary:
As per Robert's analysis:
The exception caught is OperationTimedOutException, i.e. the client driver did not hear reply from the server with the read timeout (default = 12 seconds). Checked the tserver log around the time it happened there are bunch of messages about tablet being deleted:
There was also a timeout dropping the table before the perf run. So looks like the the table drop was still in progress when the load tester proceeded. I guess we should increase the timeout for table-drop and do not ignore the exception now that we are using "DROP TABLE IF EXISTS...".
http://docs.datastax.com/en/developer/java-driver/3.2/manual/socket_options/
session.execute(new SimpleStatement("DROP TABLE IF EXISTS ...").setReadTimeoutMillis(30000));

Test Plan:
Ran
```
java -jar ~/code/yugabyte/java/yb-loadtester/target/yb-sample-apps.jar -workload CassandraKeyValue -nodes 127.0.0.1:9042,127.0.0.2:9042,127.0.0.3:9042
```
twice in a row to kick in the drop of the default table.

Reviewers: robert

Reviewed By: robert

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3019