ENG-2147: Adding support for new release format
Summary:
Adding a version.txt file to support bumping the version number for this repo.
Also added relevant changes to yb_release and itest so that we can validate and test this.

Test Plan:
Did both a yb_release and a bundle from devops.
Also ran itest and validated at least the onprem flow being able to package the builds and create/destroy universe

Reviewers: ram, jeff, karthik, bharat

Reviewed By: bharat

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D3010