Ensure we don't specify the --air_gap flag in onprem deployments.
Summary:
We would like to decouple the `local_package_path` and `air_gap` flags. We would like to
ensure that `local_package_path` is always present and we pick up these local packages everywhere.
On the other hand we'd like to customer to specify whether the install is air_gap or not when the
provider is configured. As a result, until this option is available via configure provier, we'd like
to assume air_gap is always False which would ensure itest would pass.

Test Plan: local onprem create universe.

Reviewers: ram, jeff

Reviewed By: jeff

Subscribers: jenkins-bot, yugaware, karthik

Differential Revision: https://phabricator.dev.yugabyte.com/D3077