ENG-2117: YW & UI portion of show current spot price market rate for given AZ
Summary:
As in the title, this is the YW & UI portion of the work to suggest a spot price to use. This implementation finds the max spot price for all AZs in a given region and returns 2 * that max spot price as the suggested spot price to use.

Related Devops work: https://phabricator.dev.yugabyte.com/D2927

Test Plan:
- sbt test
- go to CreateUniverse form and see that a spot price is suggested to the user:
{F5514}
{F5303}

Reviewers: bogdan, pritam

Reviewed By: pritam

Subscribers: ram, jenkins-bot, ui, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D3088