Give details of CQL queries in progress queries in the rpc dashboard
Summary:
By going to the dashboard at localhost:12001/rpcz,
it is possible to get all the cql queries
in progress for the following types of requests: query, prepare, execute and
batch. This is done by adding information to the cql call and then having
the protobuf dump code dump the queries.

Test Plan:
Run the cql loadtester in the local yugabyte cluster and go to
the dashboard to check the queries in progress.

Reviewers: kannan, robert

Reviewed By: robert

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3093