ENG-2179: Call RpcService handler directly for local call if the current thread is the worker thread.
Summary: Currently, with YB load-balalcing policy, when the tserver CQL proxy tries to call is local, the call is local without going through TCP socket. Yet, the call is made by placing the call in the tserver's inbound call queue. Same case when the tserver call finishes that the CQL call is re-queued to process the tserver response. These two queuing result in 2 thread switches. This diff now checks if the current thread is the RPC worker thread (the common case) and avoids the thread switch with a direct call. This change improves YCSB workload C latency by about 15% and throughput by around 2%.

Test Plan: Ran CassandraKeyValue read-only loadtester and verified the calls are now direct.

Reviewers: amitanand, bogdan, sergei

Reviewed By: sergei

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3079