Fix build broken by the snapshot diff
Summary:
Fix build broken by the snapshot diff. It looks like we're generating CMake target names out
of a path, and if the path contains an "@" character, we get into trouble. The root cause of this
issue is probably in something else, i.e. we're using absolute paths where we used to use relative
paths, but this fix works.

Test Plan: Jenkins

Reviewers: oleg, bogdan, akashnil, hector, karthik

Reviewed By: karthik

Subscribers: karthik, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3132