ENG-2191 Support CQL 'USING TIMESTAMP' for INSERT and UPDATES.
Summary:
This diff introduces support for the CQL USING TIMESTAMP for INSERT and UPDATE statements.
DELETE statements are currently not supported and will be implemented in a future diff.

The logic for USING TIMESTAMP is implemented at the DocDB level on a subdoc basis as
follows:

1) If the statement doesn't have user supplied timestamp, all the logic is same as before.
2) If the statement has a user supplied timestamp, look at the last value for that subdoc (including tombstones).
3) If the last value has a user supplied timestamp, use that as the existing subdoc's timestamp for comparison, otherwise just use the hybrid timestamp.
4) If the user supplied timestamp is greater than or equal to the existing subdoc's timestamp, then we apply the write, otherwise we skip it.

In addition to this, I also fixed a bug with writetime() where it was returning the int64 encoded
value of hybrid time instead of the physical component of hybrid time.

Test Plan:
1) Unit tests added.
2) jenkins

Reviewers: mihnea, rahuldesirazu, robert, akashnil, mikhail

Reviewed By: mikhail

Subscribers: neil, yql, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D3204