Affinitized Leaders
Summary:
YW and UI changes to implement the affinitized leaders feature on both the create and edit universe path. Right now, the ui is very simple, with checkboxes next to the azs in the placement info section of the form, it will need to be made to look nicer. In addition, will not build as yet since the YB changes have not yet been landed.

Test Plan: Testing by running YW on a dev server, creating/editing universes with different affinitized leader info, and using yb-admin commands to check the leader count. Open to suggestions on how to further test.

Reviewers: jeff, pritam, bharat

Reviewed By: bharat

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D3086